% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/District_Results.R
\name{District_Results}
\alias{District_Results}
\title{Results from a district}
\usage{
District_Results(state, district_no = 1)
}
\arguments{
\item{state}{The name of the state with the district you are interested in.
It must also be properly capitalized.}

\item{district_no}{The number of the district in the state. This is inputted as
a numeric (no parantheses needed).}
}
\description{
Given a state and a district in that state, this function will return all of the
election results from that district. If the state only has one district, leave the
second argument district_no blank.
}
\examples{
District_Results("New Jersey", 3)
District_Results("Alaska")
}
