% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_state_function.R
\name{gdp_state}
\alias{gdp_state}
\title{A function for state-specific per capita Gross Domestic Product data}
\usage{
gdp_state(state)
}
\arguments{
\item{state}{Name of the state you want the GDP information for. Must be capitalized.}
}
\description{
This function allows you to get state-specific GDP data.
}
\examples{
gdp_state("Virginia")
}
