% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_status_function.R
\name{gdp_status}
\alias{gdp_status}
\title{A function for state-specific, time-specific per capita Gross Domestic Product}
\usage{
gdp_status(state, year1, year2)
}
\description{
This function allows for you to see if per capita GDP increased for a given state over a given time period.
}
