% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/House_Control.R
\name{House_Control}
\alias{House_Control}
\title{Who controlled the House of Representatives in a given year?}
\usage{
House_Control(year)
}
\arguments{
\item{year}{The year that you want the information from. Must be an even year between 1998 and 2016.}
}
\description{
This function counts the number of Republicans and Democrats (and Independents) sent to Congress
in a given election year. It tells you which party controlled the House
after the election, and how many more seats that party had over the minority party.
}
\examples{
House_Control(2004)
}
