% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_prez_party.R
\name{house_prez_party}
\alias{house_prez_party}
\title{Plot House Majority with President Party}
\usage{
house_prez_party(year)
}
\arguments{
\item{year}{The year must be an even number between and including 1998 and 2016.}
}
\description{
Given a year, this function compares the dominant House party of each state versus
the party of the president.
}
\examples{
house_prez_party(2002)
}
