% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Independents_State.R
\name{Independents_State}
\alias{Independents_State}
\title{Independents sent by State}
\usage{
Independents_State(state)
}
\arguments{
\item{state}{Name of the state that you want the information for. Must be capitalized appropriately.}
}
\description{
This function counts the number of Independents a given state sent to Congress
in each election year from 1998 to 2016.
}
\examples{
Independents_State("Vermont")
}
