% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_majority_party.R
\name{pop_majority_party}
\alias{pop_majority_party}
\title{Plot Population by Representatives with Majority Party}
\usage{
pop_majority_party(year)
}
\arguments{
\item{year}{The year must be an even number between and including 1998 and 2016.}
}
\description{
Given a year, this function compares the population versus number of representatives
by state with the color of the point the state's House party majority.
}
\examples{
pop_majority_party(2002)
}
