% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Representative_Info.R
\name{Representative_Info}
\alias{Representative_Info}
\title{Information on a particular representative}
\usage{
Representative_Info(name)
}
\arguments{
\item{state}{The name of the representative, properly capitalized.#}
}
\description{
Given the name of someone who served in the US House of Representatives from
1998-2016, this function will return the years they served in the House, their
party, state and average percentage of the vote they got in the elections between
1998-2016.
}
\examples{
Reprsentative_Info("Bernie Sanders")
}
