% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reps_by_state_stacked.R
\name{reps_by_state_stacked}
\alias{reps_by_state_stacked}
\title{Plot Representatives by State}
\usage{
reps_by_state_stacked(year)
}
\arguments{
\item{year}{The year must be an even number between and including 1998 and 2016.}
}
\description{
Given a year, this function graphs the number of representatives per state 
colored by the party of the representative.
}
\examples{
reps_by_state_stacked(2002)
}
