% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/State_Party_Data.R
\name{State_Party_Data}
\alias{State_Party_Data}
\title{Number of representatives of each party sent to Congress by a given state.}
\usage{
State_Party_Data(state)
}
\arguments{
\item{state}{The name of the state, properly capitalized.#}
}
\description{
Given the name of a state, this function will use the Democrats_State, Republicans_State,
and Independents_State to create a table that tells you how many Democrats, Republicans,
Independents. The data is in a format that makes it easy to create a bar graph
representing the information, which is used in the shiny app. If a state did not
send an Independent in any year, Independents are not included in the table.
}
\examples{
State_Party_Data("Maryland")
}
